GETP-175 rename: description.response에서 description으로 이동
