+++ src/test/java/es/princip/getp/domain/people/query/presentation/description/DetailPeopleProfileResponseDescription.java
@@ -1,4 +1,4 @@
-package es.princip.getp.domain.people.query.presentation.description.response;
+package es.princip.getp.domain.people.query.presentation.description;
 
 import org.springframework.restdocs.payload.FieldDescriptor;
