+++ src/test/java/es/princip/getp/domain/people/query/presentation/description/PublicDetailPeopleResponseDescription.java
@@ -1,4 +1,4 @@
-package es.princip.getp.domain.people.query.presentation.description.response;
+package es.princip.getp.domain.people.query.presentation.description;
 
 import org.springframework.restdocs.payload.FieldDescriptor;
 
@@ -6,7 +6,7 @@ import static es.princip.getp.infra.util.FieldDescriptorHelper.getDescriptor;
 import static org.springframework.restdocs.snippet.Attributes.key;
 
 
-public class DetailPeopleResponseDescription {
+public class PublicDetailPeopleResponseDescription {
 
     public static FieldDescriptor[] description() {
         return new FieldDescriptor[] {
@@ -17,14 +17,7 @@ public class DetailPeopleResponseDescription {
                 .attributes(key("format").value("TEAM, INDIVIDUAL")),
             getDescriptor("completedProjectsCount", "완수한 프로젝트 수"),
             getDescriptor("likesCount", "받은 좋아요 수"),
-            getDescriptor("profile.introduction", "소개"),
-            getDescriptor("profile.activityArea", "활동 지역"),
-            getDescriptor("profile.techStacks[]", "기술 스택"),
-            getDescriptor("profile.education.school", "학교"),
-            getDescriptor("profile.education.major", "전공"),
-            getDescriptor("profile.hashtags[]", "해시태그"),
-            getDescriptor("profile.portfolios[].url", "포트폴리오 URL"),
-            getDescriptor("profile.portfolios[].description", "포트폴리오 설명")
+            getDescriptor("profile.hashtags[]", "해시태그")
         };
     }
 }
